Crypto.Secp256k1.Internal.Recovery

Description

Crytpographic functions related to recoverable signatures from Bitcoin’s secp256k1 library.

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

newtype RecSig

Constructors

RecSig

Fields

get :: ByteString

data CompactRecSig

Constructors

CompactRecSig

Fields

rs :: !ByteString

v :: !Word8

compactRecSig :: ByteString -> Maybe CompactRecSig

serializeCompactRecSig :: CompactRecSig -> ByteString

compactRecSigFromString :: String -> Maybe CompactRecSig

importCompactRecSig :: Ctx -> CompactRecSig -> Maybe RecSig

Parse a compact ECDSA signature (64 bytes + recovery id).

exportCompactRecSig :: Ctx -> RecSig -> CompactRecSig

Serialize an ECDSA signature in compact format (64 bytes + recovery id).

convertRecSig :: Ctx -> RecSig -> Sig

Convert a recoverable signature into a normal signature.

signRecMsg :: Ctx -> SecKey -> Msg -> RecSig

Create a recoverable ECDSA signature.

recover :: Ctx -> RecSig -> Msg -> Maybe PubKey

Recover an ECDSA public key from a signature.