Crypto.Secp256k1.Internal.RecoveryOps

Description

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

data RecSig65

ecdsaRecoverableSignatureParseCompact :: Ptr LCtx -> Ptr RecSig65 -> Ptr Compact64 -> CInt -> IO Ret

ecdsaRecoverableSignatureConvert :: Ptr LCtx -> Ptr Sig64 -> Ptr RecSig65 -> IO Ret

ecdsaRecoverableSignatureSerializeCompact :: Ptr LCtx -> Ptr Compact64 -> Ptr CInt -> Ptr RecSig65 -> IO Ret

ecdsaSignRecoverable

Arguments

:: Ptr LCtx

-> Ptr RecSig65

-> Ptr Msg32

-> Ptr SecKey32

-> FunPtr (NonceFun a)

-> Ptr a
nonce data

-> IO Ret

ecdsaRecover :: Ptr LCtx -> Ptr PubKey64 -> Ptr RecSig65 -> Ptr Msg32 -> IO Ret